requirements.txt
flask
google-generativeai
langchain
langchain-community
langchain-google-genai
chromadb
beautifulsoup4
